added helpers functions to palmsens_driver
